Improve compute shader API
Former-commit-id: 5bc036b0b3a61ff0ad0929e47505295e9ae96721

+++ Mountain/src/Mountain/rendering/render_target.cpp
@@ -50,7 +50,7 @@ void RenderTarget::Initialize(const Vector2i size, const MagnificationFilter fil
     glTextureParameteri(m_Texture, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
 
     glBindTexture(GL_TEXTURE_2D, m_Texture);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA8, size.x, size.y, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, size.x, size.y, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
     glBindTexture(GL_TEXTURE_2D, 0);
 
     // Framebuffer
